Modificado o service DeletaItemService para não permitir deletar o item se houver itemCheckList Associado a ele

+++ src/main/java/com/DougFSiva/checkMate/service/item/DeletaItemService.java
@@ -5,7 +5,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.DougFSiva.checkMate.config.imagem.ImagemConfig;
+import com.DougFSiva.checkMate.exception.ErroDeOperacaoComItemException;
 import com.DougFSiva.checkMate.model.Item;
+import com.DougFSiva.checkMate.repository.ItemCheckListRepository;
 import com.DougFSiva.checkMate.repository.ItemRepository;
 import com.DougFSiva.checkMate.service.imagem.DeletaImagemService;
 import com.DougFSiva.checkMate.util.LoggerPadrao;
@@ -18,12 +20,17 @@ public class DeletaItemService {
 
     private static final LoggerPadrao logger = new LoggerPadrao(DeletaItemService.class);
     private final ItemRepository repository;
+    private final ItemCheckListRepository itemCheckListRepository;
     private final DeletaImagemService imagemService;
     
     @Transactional
     @PreAuthorize("hasRole('ADMIN')")
 	public void deletar(Long ID) {
     	Item item = repository.findByIdOrElseThrow(ID);
+    	if (itemCheckListRepository.existsByItem(item)) {
+    		throw new ErroDeOperacaoComItemException(
+    				String.format("Não é possível deletar item com ID %s, pois há checklists associados a ele", item.getID() ));
+    	}
     	repository.delete(item);
     	deletarImagem(item);
     	logger.info(String.format("Deletado item %s", item.infoParaLog()));
@@ -34,4 +41,6 @@ public class DeletaItemService {
 			imagemService.deletarImagem(item.getImagem());
 		}
 	}
+    
+
 }
